Re-implement StorageException
