+++ src/Exception/ModifyReadOnlyEntityException.php
@@ -6,9 +6,6 @@ namespace App\Exception;
 
 use App\ErrorResponse\ModifyReadOnlyEntityErrorInterface;
 
-/**
- * @phpstan-import-type SerializedModifyReadOnlyEntityError from ModifyReadOnlyEntityErrorInterface
- */
 class ModifyReadOnlyEntityException extends FooException
 {
     public function __construct(ModifyReadOnlyEntityErrorInterface $error)
